Clean up inline helper a bit
